move get_exports_by_application to analytics
