+++ corehq/apps/app_manager/tests/test_analytics.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
-from corehq.apps.app_manager.dbaccessors import get_exports_by_application, \
-    update_analytics_indexes
+from corehq.apps.app_manager.analytics import update_analytics_indexes, \
+    get_exports_by_application
 from corehq.apps.app_manager.models import Application, Module
